refactor: move ephemeral cluster nginx code to a dedicated file
For better modularity and readability.
